+++ src/routines/estimation/qr.rs
